15_WorkingWithSwaggerOpenAPI/rest-with-spring-boo-and-java-erudio/src/test/java/br/com/engel/unittests/mockito/services/PersonServicesTest.java
package br.com.engel.unittests.mockito.services;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.Mockito.when;

import java.util.List;
import java.util.Optional;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.TestInstance;
import org.junit.jupiter.api.TestInstance.Lifecycle;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import br.com.engel.data.vo.v1.PersonVO;
import br.com.engel.exceptions.RequiredObjectIsNullException;
import br.com.engel.model.Person;
import br.com.engel.repositories.PersonRepository;
import br.com.engel.services.PersonServices;
import br.com.engel.unittests.mapper.mocks.MockPerson;

@TestInstance(Lifecycle.PER_CLASS)
@ExtendWith(MockitoExtension.class)
class PersonServicesTest {

	MockPerson input;

	@InjectMocks
	private PersonServices service;

	@Mock
	PersonRepository repository;

	@BeforeEach
	void setUpMocks() throws Exception {
		input = new MockPerson();
		MockitoAnnotations.openMocks(this);
	}

	@Test
	void testFindById() {
		Person entity = input.mockEntity(1);
		entity.setId(1L);

		when(repository.findById(1L)).thenReturn(Optional.of(entity));
		PersonVO result = service.findById(1L);
		assertNotNull(result);
		assertNotNull(result.getKey());
		assertNotNull(result.getLinks());
		assertTrue(result.toString().contains("links: [</api/person/v1/1>;rel=\"self\"]"));
		assertEquals("Addres Test1", result.getAddress());
		assertEquals("First Name Test1", result.getFirstName());
		assertEquals("Last Name Test1", result.getLastName());
		assertEquals("Female", result.getGender());
	}

	@Test
	void testCreate() throws CloneNotSupportedException {
		Person entity = input.mockEntity(1);
		entity.setId(null);
		Person persisted = entity.clone();
		persisted.setId(1L);

		PersonVO vo = input.mockVO(1);
		vo.setKey(null);

		when(repository.save(entity)).thenReturn(persisted);
		PersonVO result = service.create(vo);
		assertNotNull(result);
		assertNotNull(result.getKey());
		assertNotNull(result.getLinks());
		assertTrue(result.toString().contains("links: [</api/person/v1/1>;rel=\"self\"]"));
		assertEquals("Addres Test1", result.getAddress());
		assertEquals("First Name Test1", result.getFirstName());
		assertEquals("Last Name Test1", result.getLastName());
		assertEquals("Female", result.getGender());
	}

	@Test
	void testCreateWithNullPerson() throws CloneNotSupportedException {
		Exception exception = assertThrows(RequiredObjectIsNullException.class, () -> {
			service.create(null);
		});

		String expectedMessage = "It is not allowed to persist a null object";
		String actualMessage = exception.getMessage();
		assertTrue(actualMessage.contains(expectedMessage));
	}

	@Test
	void testUpdate() throws CloneNotSupportedException {
		Person entity = input.mockEntity(1);
		entity.setId(1L);
		Person persisted = entity.clone();
		persisted.setId(1L);

		PersonVO vo = input.mockVO(1);
		vo.setKey(1L);

		when(repository.findById(1L)).thenReturn(Optional.of(entity));
		when(repository.save(entity)).thenReturn(persisted);

		PersonVO result = service.update(vo);

		assertNotNull(result);
		assertNotNull(result.getKey());
		assertNotNull(result.getLinks());
		assertTrue(result.toString().contains("links: [</api/person/v1/1>;rel=\"self\"]"));
		assertEquals("Addres Test1", result.getAddress());
		assertEquals("First Name Test1", result.getFirstName());
		assertEquals("Last Name Test1", result.getLastName());
		assertEquals("Female", result.getGender());
	}

	@Test
	void testUpdateWithNullPerson() throws CloneNotSupportedException {
		Exception exception = assertThrows(RequiredObjectIsNullException.class, () -> {
			service.update(null);
		});

		String expectedMessage = "It is not allowed to persist a null object";
		String actualMessage = exception.getMessage();
		assertTrue(actualMessage.contains(expectedMessage));
	}

	@Test
	void testDelete() {
		Person entity = input.mockEntity(1);
		entity.setId(1L);

		when(repository.findById(1L)).thenReturn(Optional.of(entity));
		service.delete(1L);
	}

	@Test
	void testFindAll() {
		List<Person> entities = input.mockEntityList();

		when(repository.findAll()).thenReturn(entities);
		List<PersonVO> people = service.findAll();

		assertNotNull(people);
		assertEquals(14, people.size());

		for (int i =0; i < people.size(); i++) {
			assertNotNull(people.get(i));
			assertNotNull(people.get(i).getKey());
			assertNotNull(people.get(i).getLinks());
			assertTrue(people.get(i).toString().contains("links: [</api/person/v1/"+ i + ">;rel=\"self\"]"));
			assertEquals("Addres Test" + i, people.get(i).getAddress());
			assertEquals("First Name Test" + i, people.get(i).getFirstName());
			assertEquals("Last Name Test" + i, people.get(i).getLastName());
			assertEquals(((i % 2)==0) ? "Male" : "Female", people.get(i).getGender());
		}
	}
}
